+++ lib/screens/signal/signal_calendar.dart
@@ -1,28 +1,38 @@
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:periodik/models/point.dart';
-import 'package:periodik/providers/points_provider.dart';
-import 'package:periodik/screens/point/delete_point.dart';
-import 'package:periodik/screens/point/point_dialog.dart';
-import 'package:periodik/utils/date_time.dart';
-import 'package:periodik/widgets/calendar/calendar_day.dart';
 
+import '../../models/point.dart';
+import '../../providers/points_provider.dart';
+import '../../utils/date_time.dart';
+import '../../utils/hero_tags.dart';
 import '../../widgets/calendar/calendar.dart';
+import '../../widgets/calendar/calendar_day.dart';
+import '../point/delete_point.dart';
+import '../point/point_dialog.dart';
+import 'signal_screen_shell.dart';
+import 'signal_view.dart';
 
-class SignalCalendar extends ConsumerWidget {
+class SignalCalendar extends ConsumerStatefulWidget {
   const SignalCalendar({
     super.key,
     required this.id,
   });
 
   final String id;
+  @override
+  ConsumerState<ConsumerStatefulWidget> createState() =>
+      __SignalCalendarState();
+}
+
+class __SignalCalendarState extends ConsumerState<SignalCalendar> {
+  final _controller = CalendarController();
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) {
+  Widget build(BuildContext context) {
     final points = ref
             .watch(
-              pointsProvider(id),
+              pointsProvider(widget.id),
             )
             .asData
             ?.value ??
@@ -47,7 +57,7 @@ class SignalCalendar extends ConsumerWidget {
         onPressed: () {
           PointDialog.show(
             context: context,
-            signalId: id,
+            signalId: widget.id,
             point: point ??
                 Point(
                   id: '',
@@ -58,15 +68,27 @@ class SignalCalendar extends ConsumerWidget {
         onLongPressed: point != null
             ? () => deletePoint(
                   context: context,
-                  signalId: id,
+                  signalId: widget.id,
                   pointId: point.id,
                 )
             : null,
       );
     }
 
-    return Calendar(
-      builder: builder,
+    return SignalScreenShell(
+      id: widget.id,
+      view: SignalView.calendar,
+      floatingActionButton: FloatingActionButton(
+        heroTag: const ResetViewHeroTag(),
+        onPressed: () {
+          _controller.resetView();
+        },
+        child: const Icon(Icons.my_location),
+      ),
+      child: Calendar(
+        controller: _controller,
+        builder: builder,
+      ),
     );
   }
 }