+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Initializes the calendar with the current day around the middle of the screen.
 - Makes the signal names more compact in the signals calendar screen.
 - Redirects from the signals calendar screen to the signal points screen when clicking on a signal name.
+- Adds a button to reset the calendar view.
 
 ## 0.16.0